+++ src/lc3/mod.rs
@@ -1,3 +1,4 @@
 pub mod bytes;
+pub mod memory;
 pub mod opcode;
 pub mod vm;
